+++ QuanLyNhaKho/BUS/Bus.cs
@@ -32,5 +32,54 @@ namespace QuanLyNhaKho.BUS
             }
             return false;
         }
+
+        public static DataTable getListPhieuXuat()
+        {
+            return Dao.getListPhieuXuat();
+        }
+        public static DataTable SearchCuaHang(string maCH)
+        {
+            return Dao.SearchCuaHang(maCH);
+        }
+        public static string getTenCuaHang(string maCH)
+        {
+            DataTable dt = SearchCuaHang(maCH);
+            string name = dt.Rows[0]["TenCH"].ToString();
+            return name;
+        }
+        public static string getDiaChiCuaHang(string maCH)
+        {
+            DataTable dt = SearchCuaHang(maCH);
+            string name = dt.Rows[0]["DiaChi"].ToString();
+            return name;
+        }
+        public static string getTenNhanVien(string maNV)
+        {
+            DataTable dt = Dao.SearchNhanVien(maNV);
+            string name = dt.Rows[0]["HoTen"].ToString();
+            return name;
+        }
+        public static string getMaPXNext()
+        {
+            DataTable dt = Dao.getMaPXNext();
+            string maPX = dt.Rows[0][0].ToString();
+            return maPX;
+        }
+        public static DataTable getChiTietPhieuXuat(string maPX)
+        {
+            return Dao.getChiTietPhieuXuat(maPX);
+        }
+        public static DataTable SearchHangHoa(string maHH)
+        {
+            return Dao.SearchHangHoa(maHH);
+        }
+        public static  int AddChiTietPhieuXuat(ChiTietPhieuXuat chiTiet)
+        {
+            return Dao.AddChitietPhieuXuat(chiTiet);
+        }
+        public static int AddPhieuXuat(PhieuXuat phieuXuat)
+        {
+            return Dao.AddPhieuXuat(phieuXuat);
+        }
     }
 }