+++ QuanLyNhaKho/GUI/Main.Designer.cs
@@ -53,13 +53,14 @@
             this.btn_HangTonKho = new System.Windows.Forms.ToolStripMenuItem();
             this.menuStrip_TroGiup = new System.Windows.Forms.ToolStripMenuItem();
             this.panel2 = new System.Windows.Forms.Panel();
+            this.tabControlHome = new System.Windows.Forms.TabControl();
             this.label2 = new System.Windows.Forms.Label();
-            this.btn_Them = new System.Windows.Forms.ToolStripLabel();
-            this.toolStripLabel1 = new System.Windows.Forms.ToolStripLabel();
-            this.toolStripLabel2 = new System.Windows.Forms.ToolStripLabel();
-            this.toolStripLabel3 = new System.Windows.Forms.ToolStripLabel();
-            this.toolStripLabel4 = new System.Windows.Forms.ToolStripLabel();
-            this.toolStripLabel5 = new System.Windows.Forms.ToolStripLabel();
+            this.btnAdd = new System.Windows.Forms.ToolStripLabel();
+            this.btnEdit = new System.Windows.Forms.ToolStripLabel();
+            this.btnDelete = new System.Windows.Forms.ToolStripLabel();
+            this.btnSave = new System.Windows.Forms.ToolStripLabel();
+            this.btnCancel = new System.Windows.Forms.ToolStripLabel();
+            this.btnUpdate = new System.Windows.Forms.ToolStripLabel();
             this.toolStripLabel6 = new System.Windows.Forms.ToolStripLabel();
             this.panel1.SuspendLayout();
             this.toolStrip1.SuspendLayout();
@@ -81,14 +82,14 @@
             // 
             this.toolStrip1.Dock = System.Windows.Forms.DockStyle.Fill;
             this.toolStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
-            this.btn_Them,
-            this.toolStripLabel1,
-            this.toolStripLabel2,
+            this.btnAdd,
+            this.btnEdit,
+            this.btnDelete,
             this.toolStripSeparator1,
-            this.toolStripLabel3,
-            this.toolStripLabel4,
+            this.btnSave,
+            this.btnCancel,
             this.toolStripSeparator2,
-            this.toolStripLabel5,
+            this.btnUpdate,
             this.toolStripSeparator3,
             this.toolStripLabel6});
             this.toolStrip1.Location = new System.Drawing.Point(0, 24);
@@ -100,17 +101,17 @@
             // toolStripSeparator1
             // 
             this.toolStripSeparator1.Name = "toolStripSeparator1";
-            this.toolStripSeparator1.Size = new System.Drawing.Size(6, 50);
+            this.toolStripSeparator1.Size = new System.Drawing.Size(6, 29);
             // 
             // toolStripSeparator2
             // 
             this.toolStripSeparator2.Name = "toolStripSeparator2";
-            this.toolStripSeparator2.Size = new System.Drawing.Size(6, 50);
+            this.toolStripSeparator2.Size = new System.Drawing.Size(6, 29);
             // 
             // toolStripSeparator3
             // 
             this.toolStripSeparator3.Name = "toolStripSeparator3";
-            this.toolStripSeparator3.Size = new System.Drawing.Size(6, 50);
+            this.toolStripSeparator3.Size = new System.Drawing.Size(6, 29);
             // 
             // menuStrip1
             // 
@@ -190,6 +191,7 @@
             this.btn_PhieuXuat.Name = "btn_PhieuXuat";
             this.btn_PhieuXuat.Size = new System.Drawing.Size(152, 22);
             this.btn_PhieuXuat.Text = "Phiếu Xuất";
+            this.btn_PhieuXuat.Click += new System.EventHandler(this.btn_PhieuXuat_Click);
             // 
             // btn_NhaCungCap
             // 
@@ -253,12 +255,25 @@
             // 
             // panel2
             // 
+            this.panel2.Controls.Add(this.tabControlHome);
             this.panel2.Controls.Add(this.label2);
             this.panel2.Dock = System.Windows.Forms.DockStyle.Fill;
             this.panel2.Location = new System.Drawing.Point(0, 74);
             this.panel2.Name = "panel2";
             this.panel2.Size = new System.Drawing.Size(1013, 411);
             this.panel2.TabIndex = 1;
+            // 
+            // tabControlHome
+            // 
+            this.tabControlHome.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.tabControlHome.DrawMode = System.Windows.Forms.TabDrawMode.OwnerDrawFixed;
+            this.tabControlHome.Location = new System.Drawing.Point(0, 0);
+            this.tabControlHome.Name = "tabControlHome";
+            this.tabControlHome.SelectedIndex = 0;
+            this.tabControlHome.Size = new System.Drawing.Size(1248, 432);
+            this.tabControlHome.TabIndex = 1;
+            this.tabControlHome.DrawItem += new System.Windows.Forms.DrawItemEventHandler(this.tabControlHome_DrawItem);
+            this.tabControlHome.MouseDown += new System.Windows.Forms.MouseEventHandler(this.tabControlHome_MouseDown);
             // 
             // label2
             // 
@@ -270,53 +285,54 @@
             this.label2.TabIndex = 0;
             this.label2.Text = "TRANG CHỦ";
             // 
-            // btn_Them
+            // btnAdd
             // 
-            this.btn_Them.Image = global::QuanLyNhaKho.Properties.Resources.add;
-            this.btn_Them.Name = "btn_Them";
-            this.btn_Them.Size = new System.Drawing.Size(54, 47);
-            this.btn_Them.Text = "Thêm";
+            this.btnAdd.Image = global::QuanLyNhaKho.Properties.Resources.add;
+            this.btnAdd.Name = "btnAdd";
+            this.btnAdd.Size = new System.Drawing.Size(54, 26);
+            this.btnAdd.Text = "Thêm";
+            this.btnAdd.Click += new System.EventHandler(this.btnAdd_Click);
             // 
-            // toolStripLabel1
+            // btnEdit
             // 
-            this.toolStripLabel1.Image = global::QuanLyNhaKho.Properties.Resources.edit;
-            this.toolStripLabel1.Name = "toolStripLabel1";
-            this.toolStripLabel1.Size = new System.Drawing.Size(42, 47);
-            this.toolStripLabel1.Text = "Sửa";
+            this.btnEdit.Image = global::QuanLyNhaKho.Properties.Resources.edit;
+            this.btnEdit.Name = "btnEdit";
+            this.btnEdit.Size = new System.Drawing.Size(42, 26);
+            this.btnEdit.Text = "Sửa";
             // 
-            // toolStripLabel2
+            // btnDelete
             // 
-            this.toolStripLabel2.Image = global::QuanLyNhaKho.Properties.Resources.delete;
-            this.toolStripLabel2.Name = "toolStripLabel2";
-            this.toolStripLabel2.Size = new System.Drawing.Size(43, 47);
-            this.toolStripLabel2.Text = "Xóa";
+            this.btnDelete.Image = global::QuanLyNhaKho.Properties.Resources.delete;
+            this.btnDelete.Name = "btnDelete";
+            this.btnDelete.Size = new System.Drawing.Size(43, 26);
+            this.btnDelete.Text = "Xóa";
             // 
-            // toolStripLabel3
+            // btnSave
             // 
-            this.toolStripLabel3.Image = global::QuanLyNhaKho.Properties.Resources.save1;
-            this.toolStripLabel3.Name = "toolStripLabel3";
-            this.toolStripLabel3.Size = new System.Drawing.Size(43, 47);
-            this.toolStripLabel3.Text = "Lưu";
+            this.btnSave.Image = global::QuanLyNhaKho.Properties.Resources.save1;
+            this.btnSave.Name = "btnSave";
+            this.btnSave.Size = new System.Drawing.Size(43, 26);
+            this.btnSave.Text = "Lưu";
             // 
-            // toolStripLabel4
+            // btnCancel
             // 
-            this.toolStripLabel4.Image = global::QuanLyNhaKho.Properties.Resources.cen1;
-            this.toolStripLabel4.Name = "toolStripLabel4";
-            this.toolStripLabel4.Size = new System.Drawing.Size(45, 47);
-            this.toolStripLabel4.Text = "Hủy";
+            this.btnCancel.Image = global::QuanLyNhaKho.Properties.Resources.cen1;
+            this.btnCancel.Name = "btnCancel";
+            this.btnCancel.Size = new System.Drawing.Size(45, 26);
+            this.btnCancel.Text = "Hủy";
             // 
-            // toolStripLabel5
+            // btnUpdate
             // 
-            this.toolStripLabel5.Image = global::QuanLyNhaKho.Properties.Resources.refresh;
-            this.toolStripLabel5.Name = "toolStripLabel5";
-            this.toolStripLabel5.Size = new System.Drawing.Size(71, 47);
-            this.toolStripLabel5.Text = "Cập nhật";
+            this.btnUpdate.Image = global::QuanLyNhaKho.Properties.Resources.refresh;
+            this.btnUpdate.Name = "btnUpdate";
+            this.btnUpdate.Size = new System.Drawing.Size(71, 26);
+            this.btnUpdate.Text = "Cập nhật";
             // 
             // toolStripLabel6
             // 
             this.toolStripLabel6.Image = global::QuanLyNhaKho.Properties.Resources.iconprint;
             this.toolStripLabel6.Name = "toolStripLabel6";
-            this.toolStripLabel6.Size = new System.Drawing.Size(33, 47);
+            this.toolStripLabel6.Size = new System.Drawing.Size(33, 26);
             this.toolStripLabel6.Text = "In";
             // 
             // toolstrip_Them
@@ -364,20 +380,21 @@
         private System.Windows.Forms.ToolStripMenuItem menuStrip_TroGiup;
         private System.Windows.Forms.Label label2;
         private System.Windows.Forms.ToolStrip toolStrip1;
-        private System.Windows.Forms.ToolStripLabel btn_Them;
-        private System.Windows.Forms.ToolStripLabel toolStripLabel1;
-        private System.Windows.Forms.ToolStripLabel toolStripLabel2;
+        private System.Windows.Forms.ToolStripLabel btnAdd;
+        private System.Windows.Forms.ToolStripLabel btnEdit;
+        private System.Windows.Forms.ToolStripLabel btnDelete;
         private System.Windows.Forms.ToolStripSeparator toolStripSeparator1;
-        private System.Windows.Forms.ToolStripLabel toolStripLabel3;
-        private System.Windows.Forms.ToolStripLabel toolStripLabel4;
+        private System.Windows.Forms.ToolStripLabel btnSave;
+        private System.Windows.Forms.ToolStripLabel btnCancel;
         private System.Windows.Forms.ToolStripSeparator toolStripSeparator2;
-        private System.Windows.Forms.ToolStripLabel toolStripLabel5;
+        private System.Windows.Forms.ToolStripLabel btnUpdate;
         private System.Windows.Forms.ToolStripSeparator toolStripSeparator3;
         private System.Windows.Forms.ToolStripLabel toolStripLabel6;
         private System.Windows.Forms.ToolStripMenuItem btn_DangNhap;
         private System.Windows.Forms.ToolStripMenuItem btn_DoiMatKhau;
         private System.Windows.Forms.ToolStripMenuItem btn_NhapHang;
         private System.Windows.Forms.ToolStripMenuItem btn_XuatHang;
+        private System.Windows.Forms.TabControl tabControlHome;
     }
 }