+++ QuanLyNhaKho/GUI/UCPhieuXuat.Designer.cs
@@ -1,0 +1,286 @@
+﻿namespace QuanLyNhaKho.GUI
+{
+    partial class UCPhieuXuat
+    {
+        /// <summary> 
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary> 
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Component Designer generated code
+
+        /// <summary> 
+        /// Required method for Designer support - do not modify 
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.btnViewMore = new System.Windows.Forms.Button();
+            this.panel1 = new System.Windows.Forms.Panel();
+            this.dtgPhieuXuat = new System.Windows.Forms.DataGridView();
+            this.rtxtGhiChu = new System.Windows.Forms.RichTextBox();
+            this.txtTenNV = new System.Windows.Forms.TextBox();
+            this.txtMaNV = new System.Windows.Forms.TextBox();
+            this.label6 = new System.Windows.Forms.Label();
+            this.label5 = new System.Windows.Forms.Label();
+            this.label4 = new System.Windows.Forms.Label();
+            this.txtDiaChiCH = new System.Windows.Forms.TextBox();
+            this.txtTenCH = new System.Windows.Forms.TextBox();
+            this.label8 = new System.Windows.Forms.Label();
+            this.label7 = new System.Windows.Forms.Label();
+            this.txtMaCH = new System.Windows.Forms.TextBox();
+            this.label3 = new System.Windows.Forms.Label();
+            this.label2 = new System.Windows.Forms.Label();
+            this.label1 = new System.Windows.Forms.Label();
+            this.label9 = new System.Windows.Forms.Label();
+            this.txtMaPX = new System.Windows.Forms.TextBox();
+            this.dtpNgayXuat = new System.Windows.Forms.DateTimePicker();
+            this.panel1.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.dtgPhieuXuat)).BeginInit();
+            this.SuspendLayout();
+            // 
+            // btnViewMore
+            // 
+            this.btnViewMore.Location = new System.Drawing.Point(943, 218);
+            this.btnViewMore.Name = "btnViewMore";
+            this.btnViewMore.Size = new System.Drawing.Size(120, 42);
+            this.btnViewMore.TabIndex = 25;
+            this.btnViewMore.Text = "Xem chi tiết";
+            this.btnViewMore.UseVisualStyleBackColor = true;
+            this.btnViewMore.Click += new System.EventHandler(this.btnViewMore_Click);
+            // 
+            // panel1
+            // 
+            this.panel1.Controls.Add(this.dtgPhieuXuat);
+            this.panel1.Dock = System.Windows.Forms.DockStyle.Bottom;
+            this.panel1.Location = new System.Drawing.Point(0, 266);
+            this.panel1.Name = "panel1";
+            this.panel1.Size = new System.Drawing.Size(1168, 297);
+            this.panel1.TabIndex = 24;
+            // 
+            // dtgPhieuXuat
+            // 
+            this.dtgPhieuXuat.ColumnHeadersHeightSizeMode = System.Windows.Forms.DataGridViewColumnHeadersHeightSizeMode.AutoSize;
+            this.dtgPhieuXuat.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.dtgPhieuXuat.Location = new System.Drawing.Point(0, 0);
+            this.dtgPhieuXuat.Name = "dtgPhieuXuat";
+            this.dtgPhieuXuat.ReadOnly = true;
+            this.dtgPhieuXuat.Size = new System.Drawing.Size(1168, 297);
+            this.dtgPhieuXuat.TabIndex = 0;
+            this.dtgPhieuXuat.CellContentClick += new System.Windows.Forms.DataGridViewCellEventHandler(this.dtgPhieuXuat_CellContentClick);
+            // 
+            // rtxtGhiChu
+            // 
+            this.rtxtGhiChu.Location = new System.Drawing.Point(139, 179);
+            this.rtxtGhiChu.Name = "rtxtGhiChu";
+            this.rtxtGhiChu.Size = new System.Drawing.Size(765, 83);
+            this.rtxtGhiChu.TabIndex = 23;
+            this.rtxtGhiChu.Text = "";
+            // 
+            // txtTenNV
+            // 
+            this.txtTenNV.Enabled = false;
+            this.txtTenNV.Location = new System.Drawing.Point(943, 98);
+            this.txtTenNV.Name = "txtTenNV";
+            this.txtTenNV.Size = new System.Drawing.Size(146, 20);
+            this.txtTenNV.TabIndex = 22;
+            // 
+            // txtMaNV
+            // 
+            this.txtMaNV.Location = new System.Drawing.Point(989, 67);
+            this.txtMaNV.Name = "txtMaNV";
+            this.txtMaNV.Size = new System.Drawing.Size(100, 20);
+            this.txtMaNV.TabIndex = 21;
+            // 
+            // label6
+            // 
+            this.label6.AutoSize = true;
+            this.label6.Location = new System.Drawing.Point(855, 98);
+            this.label6.Name = "label6";
+            this.label6.Size = new System.Drawing.Size(82, 13);
+            this.label6.TabIndex = 18;
+            this.label6.Text = "Tên nhân viên :";
+            // 
+            // label5
+            // 
+            this.label5.AutoSize = true;
+            this.label5.Location = new System.Drawing.Point(855, 68);
+            this.label5.Name = "label5";
+            this.label5.Size = new System.Drawing.Size(128, 13);
+            this.label5.TabIndex = 19;
+            this.label5.Text = "Mã nhân viên giao hàng :";
+            // 
+            // label4
+            // 
+            this.label4.AutoSize = true;
+            this.label4.Location = new System.Drawing.Point(462, 67);
+            this.label4.Name = "label4";
+            this.label4.Size = new System.Drawing.Size(61, 13);
+            this.label4.TabIndex = 17;
+            this.label4.Text = "Ngày xuất :";
+            // 
+            // txtDiaChiCH
+            // 
+            this.txtDiaChiCH.Enabled = false;
+            this.txtDiaChiCH.Location = new System.Drawing.Point(139, 142);
+            this.txtDiaChiCH.Name = "txtDiaChiCH";
+            this.txtDiaChiCH.Size = new System.Drawing.Size(249, 20);
+            this.txtDiaChiCH.TabIndex = 16;
+            // 
+            // txtTenCH
+            // 
+            this.txtTenCH.Enabled = false;
+            this.txtTenCH.Location = new System.Drawing.Point(140, 107);
+            this.txtTenCH.Name = "txtTenCH";
+            this.txtTenCH.Size = new System.Drawing.Size(130, 20);
+            this.txtTenCH.TabIndex = 15;
+            // 
+            // label8
+            // 
+            this.label8.AutoSize = true;
+            this.label8.Location = new System.Drawing.Point(44, 202);
+            this.label8.Name = "label8";
+            this.label8.Size = new System.Drawing.Size(50, 13);
+            this.label8.TabIndex = 12;
+            this.label8.Text = "Ghi chú :";
+            // 
+            // label7
+            // 
+            this.label7.AutoSize = true;
+            this.label7.Location = new System.Drawing.Point(40, 145);
+            this.label7.Name = "label7";
+            this.label7.Size = new System.Drawing.Size(94, 13);
+            this.label7.TabIndex = 11;
+            this.label7.Text = "Địa chỉ cửa hàng :";
+            // 
+            // txtMaCH
+            // 
+            this.txtMaCH.Location = new System.Drawing.Point(140, 76);
+            this.txtMaCH.Name = "txtMaCH";
+            this.txtMaCH.Size = new System.Drawing.Size(100, 20);
+            this.txtMaCH.TabIndex = 14;
+            // 
+            // label3
+            // 
+            this.label3.AutoSize = true;
+            this.label3.Location = new System.Drawing.Point(40, 110);
+            this.label3.Name = "label3";
+            this.label3.Size = new System.Drawing.Size(80, 13);
+            this.label3.TabIndex = 10;
+            this.label3.Text = "Tên cửa hàng :";
+            // 
+            // label2
+            // 
+            this.label2.AutoSize = true;
+            this.label2.Location = new System.Drawing.Point(44, 79);
+            this.label2.Name = "label2";
+            this.label2.Size = new System.Drawing.Size(76, 13);
+            this.label2.TabIndex = 13;
+            this.label2.Text = "Mã cửa hàng :";
+            // 
+            // label1
+            // 
+            this.label1.AutoSize = true;
+            this.label1.Font = new System.Drawing.Font("Microsoft Sans Serif", 15.75F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.label1.ForeColor = System.Drawing.Color.Firebrick;
+            this.label1.Location = new System.Drawing.Point(503, 11);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(175, 25);
+            this.label1.TabIndex = 9;
+            this.label1.Text = "Phiếu Xuất Kho";
+            // 
+            // label9
+            // 
+            this.label9.AutoSize = true;
+            this.label9.Location = new System.Drawing.Point(43, 48);
+            this.label9.Name = "label9";
+            this.label9.Size = new System.Drawing.Size(80, 13);
+            this.label9.TabIndex = 13;
+            this.label9.Text = "Mã phiếu xuất :";
+            // 
+            // txtMaPX
+            // 
+            this.txtMaPX.Location = new System.Drawing.Point(139, 45);
+            this.txtMaPX.Name = "txtMaPX";
+            this.txtMaPX.Size = new System.Drawing.Size(100, 20);
+            this.txtMaPX.TabIndex = 14;
+            // 
+            // dtpNgayXuat
+            // 
+            this.dtpNgayXuat.CustomFormat = "";
+            this.dtpNgayXuat.Format = System.Windows.Forms.DateTimePickerFormat.Custom;
+            this.dtpNgayXuat.Location = new System.Drawing.Point(541, 61);
+            this.dtpNgayXuat.Name = "dtpNgayXuat";
+            this.dtpNgayXuat.Size = new System.Drawing.Size(123, 20);
+            this.dtpNgayXuat.TabIndex = 26;
+            // 
+            // UCPhieuXuat
+            // 
+            this.AccessibleDescription = "Phiếu xuất";
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.Controls.Add(this.dtpNgayXuat);
+            this.Controls.Add(this.btnViewMore);
+            this.Controls.Add(this.panel1);
+            this.Controls.Add(this.rtxtGhiChu);
+            this.Controls.Add(this.txtTenNV);
+            this.Controls.Add(this.txtMaNV);
+            this.Controls.Add(this.label6);
+            this.Controls.Add(this.label5);
+            this.Controls.Add(this.label4);
+            this.Controls.Add(this.txtDiaChiCH);
+            this.Controls.Add(this.txtTenCH);
+            this.Controls.Add(this.label8);
+            this.Controls.Add(this.label7);
+            this.Controls.Add(this.txtMaPX);
+            this.Controls.Add(this.txtMaCH);
+            this.Controls.Add(this.label9);
+            this.Controls.Add(this.label3);
+            this.Controls.Add(this.label2);
+            this.Controls.Add(this.label1);
+            this.Name = "UCPhieuXuat";
+            this.Size = new System.Drawing.Size(1168, 563);
+            this.panel1.ResumeLayout(false);
+            ((System.ComponentModel.ISupportInitialize)(this.dtgPhieuXuat)).EndInit();
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private System.Windows.Forms.Button btnViewMore;
+        private System.Windows.Forms.Panel panel1;
+        private System.Windows.Forms.DataGridView dtgPhieuXuat;
+        private System.Windows.Forms.RichTextBox rtxtGhiChu;
+        private System.Windows.Forms.TextBox txtTenNV;
+        private System.Windows.Forms.TextBox txtMaNV;
+        private System.Windows.Forms.Label label6;
+        private System.Windows.Forms.Label label5;
+        private System.Windows.Forms.Label label4;
+        private System.Windows.Forms.TextBox txtDiaChiCH;
+        private System.Windows.Forms.TextBox txtTenCH;
+        private System.Windows.Forms.Label label8;
+        private System.Windows.Forms.Label label7;
+        private System.Windows.Forms.TextBox txtMaCH;
+        private System.Windows.Forms.Label label3;
+        private System.Windows.Forms.Label label2;
+        private System.Windows.Forms.Label label1;
+        private System.Windows.Forms.Label label9;
+        private System.Windows.Forms.TextBox txtMaPX;
+        private System.Windows.Forms.DateTimePicker dtpNgayXuat;
+    }
+}